+++ previous-versions/commerce/pkg/models/shared/companyinfo.go
@@ -61,8 +61,6 @@ func (o *Weblink) GetURL() *string {
 
 // CompanyInfo - In the Codat system, company profile includes standard commercial details about
 // a linked company, such as their address, phone number, and company registration.
-//
-// Explore our [data coverage](https://knowledge.codat.io/supported-features/commerce?view=tab-by-data-type&dataType=companyInfo) for this data type.
 type CompanyInfo struct {
 	// The available and current cash balances for the company's accounts
 	AccountBalances []AccountBalance `json:"accountBalances,omitempty"`