+++ previous-versions/commerce/pkg/models/shared/customer.go
@@ -3,8 +3,6 @@
 package shared
 
 // Customer - When a customer places an order with the connected commerce store their details are added to the Customers dataset. You can use the data from the Customers endpoints to calculate key metrics, such as customer churn.
-//
-// Explore our [data coverage](https://knowledge.codat.io/supported-features/commerce?view=tab-by-data-type&dataType=commerce-customers) for this data type.
 type Customer struct {
 	// Addresses of the customer
 	Addresses []Address `json:"addresses,omitempty"`