+++ previous-versions/commerce/pkg/models/shared/productcategory.go
@@ -3,8 +3,6 @@
 package shared
 
 // ProductCategory - Product categories are used to classify a group of products together, either by type (e.g. "Furniture"), or sometimes by tax profile.
-//
-// Explore our [data coverage](https://knowledge.codat.io/supported-features/commerce?view=tab-by-data-type&dataType=commerce-productCategories) for this data type.
 type ProductCategory struct {
 	// A collection of parent product categories implicitly ordered with the immediate parent last in the list.
 	AncestorRefs []RecordRef `json:"ancestorRefs,omitempty"`