+++ previous-versions/commerce/pkg/models/shared/productinventorylocation.go
@@ -3,7 +3,7 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/commerce/pkg/utils"
+	"github.com/codatio/client-sdk-go/previous-versions/commerce/v3/pkg/utils"
 	"github.com/ericlagergren/decimal"
 )